requirements.txt
requests
aiocron==1.6
async-files==0.4
beautifulsoup4==4.9.3
httpx[http2]==0.23.0
humanize==4.3.0
langs==0.1.1
meval==2.5
pendulum==2.1.2
pyromod==1.5
python-rapidjson==1.4
PySocks==1.7.1
PyYAML==5.4.1
psutil==5.8.0
TgCrypto==1.2.2
tortoise-orm==0.17.6
telegraph==1.4.1
typing-extensions==3.10.0.0
Pyrogram==1.4.16
pytz
asyncpg
python-anilist==1.0.4




#aiocron==1.6
#async-files==0.4
#beautifulsoup4==4.9.3
#httpx[http2]==0.23.0
#humanize==4.3.0
#langs==0.1.1
#meval==2.5
#pendulum==2.1.2
#pyromod==1.5
#python-rapidjson==1.4
#PySocks==1.7.1
#PyYAML==5.4.1
#psutil==5.8.0
#TgCrypto==1.2.2
#tortoise-orm
#tortoise-orm==0.17.6
#telegraph==1.4.1
#typing-extensions==3.10.0.0
#Pyrogram==1.4.16
#pysqlite3
#db-sqlite3
#sqlite-utils
#pytz
#asyncpg
#aiomysql
